// moonbowcl/src/app/model/dto/processo.ts
/*
* SPDX-FileCopyrightText: (C) Copyright 2023 C.S.I. Piemonte
*
* SPDX-License-Identifier: EUPL-1.2 */

export class Processo {
    idProcesso: number;
    codiceProcesso: string;
    nomeProcesso: string;
    descProcesso: string;
    flagAttivo: string;
    dataUpd: Date;
    attoreUpd: string;

    constructor() {
    }
}